GUI folder split into more files to simplify reading
